test(e2e): isolate caption position regression
Keep the focused caption regression in its own spec so changed-test selection does not pull in unrelated watch-page failures.
